Add files via upload
This is a school project that has four parts, the first part asks for the width, height (in pixels) and resolution of pixels and then it tells how large they will be in inches. the second part asks for grades in each section of a class then it calculates the average grade. the third part asks for a letter then it prints out the next three letters.