Add files via upload
this is a four part project.  The first part is an adventure game. the second part is a code that asks the user what grade they are in karate, then tells them what belt they have. the third part is a program that determines weather or not a certain app will work on your device, by asking a series of questions.